pyproject.toml
[project]
name = "loclock-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi>=0.121.3",
    "passlib>=1.7.4",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyotp>=2.9.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "qrcode>=8.2",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "user-agents>=2.2.0",
    "uvicorn>=0.38.0",
]
